Changed the producer method so that it keeps track of all its processes in a vector. The processes are started in a loop, and it polls the message queue for each thread.

+++ pde/approx/butterflyVWaspPDE/numericaltrials.h
@@ -43,7 +43,7 @@ public:
             double dt, int maxTimeLupe,
             int legendrePolyDegree,
             double maxDeltaNorm, int maxNewtonSteps,
-            int skipPrint, int msgID, int which);
+            int skipPrint, int msgID,unsigned long which);
 
 protected:
 
@@ -52,7 +52,7 @@ private:
     struct MaxMinBuffer
     {
       long mtype;
-      int which;
+      unsigned long which;
       double maxButterfly;
       double minButterfly;
       double maxWasp;
